requirements.txt
#
####### example-requirements.txt #######
#
###### Requirements without Version Specifiers ######
pysub
RPI.GPIO
adafruit-blinka
adafruit-circuitpython-neopixel
picamera
quart
pydub

#-r other-requirements.txt
PyPubSub